+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/loader/jdbc/ms/MsRolesReader.java
@@ -68,26 +68,6 @@ public class MsRolesReader extends AbstractStatementReader {
         .where("res.name != N'public'");
     }
 
-    @Override
-    protected void addMsPriviligesPart(QueryBuilder builder) {
-        String acl = """
-                CROSS APPLY (
-                  SELECT * FROM (
-                    SELECT
-                      perm.state_desc AS sd,
-                      perm.permission_name AS pn,
-                      roleprinc.name AS r
-                    FROM sys.database_principals roleprinc WITH (NOLOCK)
-                    JOIN sys.database_permissions perm WITH (NOLOCK) ON perm.grantee_principal_id = roleprinc.principal_id
-                    WHERE major_id = res.principal_id AND perm.class = 4
-                  ) cc\s
-                  FOR XML RAW, ROOT
-                ) aa (acl)""";
-
-        builder.column("aa.acl");
-        builder.join(acl);
-    }
-
     private void addMsGroupsPart(QueryBuilder builder) {
         String groups = """
                 CROSS APPLY (
@@ -102,5 +82,10 @@ public class MsRolesReader extends AbstractStatementReader {
 
         builder.column("cc.groups");
         builder.join(groups);
+    }
+
+    @Override
+    protected String getFormattedMsPriviliges() {
+        return MS_PRIVILIGES_JOIN.formatted("", "principal_id", 4);
     }
 }